+++ Eflatun.SceneReference/Packages/com.eflatun.scenereference/Runtime/InvalidSceneReferenceException.cs
@@ -5,9 +5,9 @@ using JetBrains.Annotations;
 namespace Eflatun.SceneReference
 {
     /// <summary>
-    /// Thrown if a SceneReference is invalid. This can happen for these reasons:
+    /// Thrown if a <see cref="SceneReference"/> is invalid. This can happen for these reasons:
     /// <list type="number">
-    /// <item>The SceneReference is not assigned anything, or is assigned an invalid or a null scene. To fix this, make sure the SceneReference is assigned a valid scene asset.</item>
+    /// <item>The <see cref="SceneReference"/> is assigned an invalid scene, or the assigned asset is not a scene. To fix this, make sure the <see cref="SceneReference"/> is assigned a valid scene asset.</item>
     /// <item>The scene GUID to path map is outdated. To fix this, you can either manually run the map generator, or enable all generation triggers. It is highly recommended to keep all the generation triggers enabled.</item>
     /// </list>
     /// </summary>
@@ -16,8 +16,8 @@ namespace Eflatun.SceneReference
     public class InvalidSceneReferenceException : SceneReferenceException
     {
         private const string ExceptionMessage =
-            "A SceneReference is invalid. This can happen for these reasons:"
-            + "\n1. The SceneReference is not assigned anything, or is assigned an invalid or a null scene. To fix this, make sure to assign a valid scene to the SceneReference."
+            "The SceneReference is invalid. This can happen for these reasons:"
+            + "\n1. The SceneReference is assigned an invalid scene, or the assigned asset is not a scene. To fix this, make sure to assign a valid scene to the SceneReference."
             + "\n2. The scene GUID to path map is outdated. To fix this, you can either manually run the generator, or enable generation triggers. It is highly recommended to keep all the generation triggers enabled.";
 
         internal InvalidSceneReferenceException() : base(ExceptionMessage)