+++ Eflatun.SceneReference/Packages/com.eflatun.scenereference/Runtime/SceneReference.cs
@@ -12,8 +12,11 @@ namespace Eflatun.SceneReference
     [Serializable]
     public class SceneReference
     {
-        [SerializeField] internal UnityEngine.Object sceneAsset;
-        [SerializeField] internal string sceneAssetGuidHex;
+        // An invalid GUID hex contains all zeros. A GUID hex has 32 chars.
+        private const string AllZeroGuidHex = "00000000000000000000000000000000";
+
+        [SerializeField] internal UnityEngine.Object sceneAsset = null;
+        [SerializeField] internal string sceneAssetGuidHex = AllZeroGuidHex;
 
         /// <summary>
         /// GUID of the scene asset in hex format. 
@@ -23,20 +26,31 @@ namespace Eflatun.SceneReference
         /// <summary>
         /// Path to the scene asset.
         /// </summary>
-        /// <exception cref="InvalidSceneReferenceException">
-        /// Throws if <see cref="IsInSceneGuidToPathMap"/> is <c>false</c>.
-        /// </exception>
-        public string Path =>
-            SceneGuidToPathMapProvider.SceneGuidToPathMap.TryGetValue(sceneAssetGuidHex, out var scenePath)
-                ? scenePath
-                : throw new InvalidSceneReferenceException();
+        /// <exception cref="EmptySceneReferenceException">Throws if <see cref="HasValue"/> is <c>false</c>.</exception>
+        /// <exception cref="InvalidSceneReferenceException">Throws if <see cref="IsInSceneGuidToPathMap"/> is <c>false</c>.</exception>
+        public string Path
+        {
+            get
+            {
+                if (!HasValue)
+                {
+                    throw new EmptySceneReferenceException();
+                }
+
+                if (!SceneGuidToPathMapProvider.SceneGuidToPathMap.TryGetValue(sceneAssetGuidHex, out var scenePath))
+                {
+                    throw new InvalidSceneReferenceException();
+                }
+
+                return scenePath;
+            }
+        }
 
         /// <summary>
         /// Build index of the scene.
         /// </summary>
-        /// <exception cref="InvalidSceneReferenceException">
-        /// Throws if <see cref="IsInSceneGuidToPathMap"/> is <c>false</c>.
-        /// </exception>
+        /// <exception cref="EmptySceneReferenceException">Throws if <see cref="HasValue"/> is <c>false</c>.</exception>
+        /// <exception cref="InvalidSceneReferenceException">Throws if <see cref="IsInSceneGuidToPathMap"/> is <c>false</c>.</exception>
         /// <remarks>
         /// This property will return <c>-1</c> if <see cref="IsInBuildAndEnabled"/> is <c>false</c>.
         /// </remarks>
@@ -45,47 +59,88 @@ namespace Eflatun.SceneReference
         /// <summary>
         /// Name of the scene asset.
         /// </summary>
-        /// <exception cref="InvalidSceneReferenceException">
-        /// Throws if <see cref="IsInSceneGuidToPathMap"/> is <c>false</c>.
-        /// </exception>
+        /// <exception cref="EmptySceneReferenceException">Throws if <see cref="HasValue"/> is <c>false</c>.</exception>
+        /// <exception cref="InvalidSceneReferenceException">Throws if <see cref="IsInSceneGuidToPathMap"/> is <c>false</c>.</exception>
         public string Name => System.IO.Path.GetFileNameWithoutExtension(Path);
 
         /// <summary>
         /// The <see cref="Scene"/> struct for this scene. Only valid if the scene is currently loaded.
         /// </summary>
-        /// <exception cref="InvalidSceneReferenceException">
-        /// Throws if <see cref="IsInSceneGuidToPathMap"/> is <c>false</c>.
-        /// </exception>
+        /// <exception cref="EmptySceneReferenceException">Throws if <see cref="HasValue"/> is <c>false</c>.</exception>
+        /// <exception cref="InvalidSceneReferenceException">Throws if <see cref="IsInSceneGuidToPathMap"/> is <c>false</c>.</exception>
         public Scene LoadedScene => SceneManager.GetSceneByPath(Path);
+
+        /// <summary>
+        /// Is this <see cref="SceneReference"/> assigned something?
+        /// </summary>
+        /// <remarks>
+        /// This property alone does not communicate if this <see cref="SceneReference"/> is safe to use. Check <see cref="IsSafeToUse"/> for that purpose.
+        /// </remarks>
+        /// <seealso cref="IsInSceneGuidToPathMap"/>
+        /// <seealso cref="IsInBuildAndEnabled"/>
+        /// <seealso cref="IsSafeToUse"/>
+        public bool HasValue
+        {
+            get
+            {
+                if (string.IsNullOrWhiteSpace(AssetGuidHex))
+                {
+                    // internal exceptions should not be documented as part of the public API
+                    throw SceneReferenceInternalException.AssetGuidHexNullOrWhiteSpace("54783205", AssetGuidHex);
+                }
+
+                return AssetGuidHex != AllZeroGuidHex;
+            }
+        }
 
         /// <summary>
         /// Does the scene GUID to path map contain the scene?
         /// </summary>
+        /// <exception cref="EmptySceneReferenceException">Throws if <see cref="HasValue"/> is <c>false</c>.</exception>
         /// <remarks>
         /// This property alone does not communicate if this <see cref="SceneReference"/> is safe to use. Check <see cref="IsSafeToUse"/> for that purpose.
         /// </remarks>
+        /// <seealso cref="HasValue"/>
         /// <seealso cref="IsInBuildAndEnabled"/>
-        public bool IsInSceneGuidToPathMap => SceneGuidToPathMapProvider.SceneGuidToPathMap.ContainsKey(AssetGuidHex);
+        /// <seealso cref="IsSafeToUse"/>
+        public bool IsInSceneGuidToPathMap
+        {
+            get
+            {
+                if (!HasValue)
+                {
+                    throw new EmptySceneReferenceException();
+                }
+                
+                return SceneGuidToPathMapProvider.SceneGuidToPathMap.ContainsKey(AssetGuidHex);
+            }
+        }
 
         /// <summary>
         /// Is the scene added and enabled in Build Settings?
         /// </summary>
-        /// <exception cref="InvalidSceneReferenceException">
-        /// Throws if <see cref="IsInSceneGuidToPathMap"/> is <c>false</c>.
-        /// </exception>
+        /// <exception cref="EmptySceneReferenceException">Throws if <see cref="HasValue"/> is <c>false</c>.</exception>
+        /// <exception cref="InvalidSceneReferenceException">Throws if <see cref="IsInSceneGuidToPathMap"/> is <c>false</c>.</exception>
         /// <remarks>
         /// This property alone does not communicate if this <see cref="SceneReference"/> is safe to use. Check <see cref="IsSafeToUse"/> for that purpose.
         /// </remarks>
+        /// <seealso cref="HasValue"/>
+        /// <seealso cref="IsInSceneGuidToPathMap"/>
+        /// <seealso cref="IsSafeToUse"/>
         public bool IsInBuildAndEnabled => BuildIndex != -1;
 
         /// <summary>
         /// Is this <see cref="SceneReference"/> safe to use?
         /// </summary>
         /// <remarks>
-        /// This property is equivalent to checking both <see cref="IsInSceneGuidToPathMap"/> and <see cref="IsInBuildAndEnabled"/>, with slightly better performance.
+        /// This property is equivalent to checking all three of <see cref="HasValue"/>, <see cref="IsInSceneGuidToPathMap"/>, and <see cref="IsInBuildAndEnabled"/>, with slightly better performance.
         /// </remarks>
+        /// <seealso cref="HasValue"/>
+        /// <seealso cref="IsInSceneGuidToPathMap"/>
+        /// <seealso cref="IsInBuildAndEnabled"/>
         public bool IsSafeToUse =>
-            SceneGuidToPathMapProvider.SceneGuidToPathMap.TryGetValue(AssetGuidHex, out var path)
+            HasValue
+            && SceneGuidToPathMapProvider.SceneGuidToPathMap.TryGetValue(AssetGuidHex, out var path)
             && SceneUtility.GetBuildIndexByScenePath(path) != -1;
     }
 }