+++ Eflatun.SceneReference/Packages/com.eflatun.scenereference/CHANGELOG.md
@@ -18,7 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   - `IsSafeToUse`
 
 ### Changed
-- Now throws more descriptive custom exceptions (`InvalidSceneAssetException` and `OutdatedSceneGuidToPathMapException`) for edge cases.
+- Now throwing more descriptive custom exceptions (`InvalidSceneAssetException` and `OutdatedSceneGuidToPathMapException`) for edge cases.
 
 ### Removed