+++ Eflatun.SceneReference/Packages/com.eflatun.scenereference/CHANGELOG.md
@@ -10,10 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Descriptive custom exceptions:
   - `SceneReferenceException`: The root class of all custom exceptions of this package.
-  - `InvalidSceneAssetException`
-  - `OutdatedSceneGuidToPathMapException`
+  - `InvalidSceneReferenceException`
 - Validation properties in `SceneReference`:
-  - `IsValidSceneAsset`
+  - `IsInSceneGuidToPathMap`
   - `IsInBuildAndEnabled`
   - `IsSafeToUse`